So I added a navigation template that I will load. Into all my HTML documents. Also I updated. The new Javascript. Load NAV.
So it handles the loading of the template and also the drop down behavior when you click it in iPhone. mode.

It was required to do it this way because if not it gets messy and doesn't work

All the other updates you see in CSS drop down and HTML dropdown is just me adding the new code and getting rid of the old one.

To reflect this change.