Set up the skeleton for my next project in Javascript.
Folder all notes is where I keep the templates that are loaded.

CSS is where I keep the styles for global effects. Like templates and. The HTML layout.

Images is where I keep the images in use.

Javascript dropdown is where the different HTML pages go for the website.

The About Me page is the About Me page.

The index is the starting page.

The license.

To load JavaScript's that are used to load templates.

Nav HTML. Used for the navigation bar.

And of course the skeleton of the reedmeat.