Added these templates. JavaScript: Errors / Throw / Try-Catch-Finally JavaScript: Scope
w3schools Is now just giving me. Different methods for Javascript and general rules that I already understand about coding.
I don't feel it's required to put it in my Javascript notes, so I think it's going to stop here. With the new templates.

Maybe I'll come back and add more templates, but for now I'm going to leave it here with the templates.

The next step is configuring all the templates in the correct format.