Added all the Different sections of HTML documents for the notes and for the drop down section.
I also moved examples of my notes to all notes so it's easy to access.

All the drop down uses a skeleton and a little different section just explaining what each one is.

Also, of course, the navigation bar HTML has been updated.
To reflect the path to the different HTML documents in drop down JavaScript.

And index has just been a little bit updated to reflect that it was 21 templates for Javascript.

Now I plan to merge this with my HTML and CSS. File on Github.