It's late, so I don't really feel like getting into the security features right now.
And also I plan to start on Typescript what I feel is a better use of my time.

I'm really not that concerned about someone hacking my website at this early stage.

Lastly, there are some unfinished. Things in the templates, but they're not really that Important for me to fix it this time.

So I stored the mall in all notes in example HTML.

And put in the readme. That's something in the future to fix so I don't forget.