Alright, I removed all the extra icon links and just use the safe Google One.
From all the HTML documents from drop down and other pages that are used.

Apparently it got carried over to all of them after I needed to use it one time for 1 template.

Where it's implemented only in that one location.

Which is CS, drop slash icons and links, HTML.

I did implement the JavaScript method and head folder that will hold the different.

Things that can be implemented into head easily if it needs required.

I didn't say point to remove it since I now implemented it and it works.

Also already checked the security feature, it's safe to use.