Move JavaScript into the JS folder.
For security reasons, I've been informed that I need to move all JavaScript.

Into that folder and not have anything in the HTML because that runs a risk.

Once again, I'm not worried about anyone hacking my website, but this is good practice. It's good to learn.

And it's very annoying, so I will probably remember this next time whenever I program with Javascript and HTML.

Or it's more advanced Language versions