RUST: Ownership & Borrowing RUST: Strings
Templates have been added, also the. AI prompt for template has also been updated.

The README was also updated to reflect the lesson I learned with AI prompts templates.

You can't expect AI to do all the heavy lifting, and it's also lazy. And bad practice.

You always have to remember that AI, while an incredibly useful tool. Is that a tool? You do not rely on it to do everything for you.

It can create great templates But those templates are always going to need touch ups and fixes because they're not going to come out perfect.