<template> Rust: Data Structures (Array, Vector, Tuple, HashMap, Structs, Enums)
Has been completed. What is the last template? Also I got rid of the example that was in my end folder. It was no longer needed.

Next I'm just going to. Add in the Rust actual code so it exists and I get some practice with it. And then I will be done with this language and moving on to the next one.