So I'm about to rearrange my files, so I'm making a backup commit in case something goes wrong.
Also I'm gonna be testing. Chat GPT.

To see if it's capable of recognizing how I change the filing system. By looking at my GitHub repo.

And then telling me quickly all the Code pass that need to be updated due to the changes I made.

If this works, then this is a very quick way to fix naming structures for files when a project gets really big.